Move file back
