+++ test/service/test_user_service.py
@@ -2,7 +2,7 @@ import logging
 from unittest import mock
 
 from azul import config
-from azul.service.responseobjects.user_service import UserService, UpdateError
+from azul.service.user_service import UserService, UpdateError
 from dynamo_test_case import DynamoTestCase
